requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# X/Twitter API
xdk>=0.1.0
requests>=2.31.0

# xAI SDK
xai-sdk>=0.1.0

# HTTP client
httpx>=0.25.0

# Web framework
flask>=3.0.0

# Supabase client
supabase>=2.3.0

